Switch to WebGPURenderer
- Updated App.tsx to use WebGPURenderer from three.js addons.
- Added @webgpu/types for type support.
- Wrapped scene components in Suspense for potential async operations.
