--- src/App.js
@@ -1,35 +1,0 @@
-import React, { useRef } from 'react'
-import { Canvas, useFrame } from '@react-three/fiber'
-import { OrbitControls } from '@react-three/drei'
-
-function Box(props) {
-  // This reference will give us direct access to the mesh
-  const meshRef = useRef()
-
-  // Rotate mesh every frame, this is outside of React's render cycle
-  useFrame(() => {
-    if (meshRef.current) {
-      meshRef.current.rotation.x += 0.01
-      meshRef.current.rotation.y += 0.01
-    }
-  })
-
-  return (
-    <mesh {...props} ref={meshRef}>
-      <boxGeometry args={[1, 1, 1]} />
-      <meshStandardMaterial color={'orange'} />
-    </mesh>
-  )
-}
-
-export default function App() {
-  return (
-    <Canvas style={{ height: '100vh', width: '100vw' }}>
-      <ambientLight />
-      <pointLight position={[10, 10, 10]} />
-      <Box position={[-1.2, 0, 0]} />
-      <Box position={[1.2, 0, 0]} />
-      <OrbitControls />
-    </Canvas>
-  )
-}
